pyproject.toml
[project]
authors = [{ name = "pablovela5620", email = "pablovela5620@gmail.com" }]
dependencies = [
    "opencv-python>=4.10.0.84,<5",
    "requests>=2.32.3,<3",
    "tqdm>=4.67.0,<5",
    "rerun-sdk",
]
description = "A rerun data loader for sequential colmap data."
name = "rerun-loader-colmap"
requires-python = ">= 3.11"
version = "0.1.0"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rerun_loader_colmap*"]

[project.scripts]
rerun-loader-colmap = "rerun_loader_colmap.main:main"

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
rerun_loader_colmap = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.dependencies]
rerun-sdk = ">=0.20.1,<0.21"
icecream = ">=2.1.3,<3"
